Add run script, add Grow hooks (commented out)
Added my run script; simply invoke `./run' as a one-stop-shop script
for compiling the project, building the docker image and container
and launching the microservice in the foreground. Press Ctrl+C to
terminate the microservice.

Type:
  ./run --help
to see a list of arguments/environment variables that the ./run script recognises.

Also added commented-out blocks for integrating Grow into a microservice.

Updated gradle wrapper to 4.8.1.