JVM remote debugging now works with Java 9/10/11+
In Java 9, there was a change introduced to the JVM remote debugger where it
will by default only listen on the loopback network interface (127.0.0.1).

The 'address=...' argument of the '-agentlib' JVM command line option now needs
to be prefixed with '*:' in order to instruct the JVM to bind the remote
debugging listening TCP socket to all network interfaces (so external debugging
clients such as Eclipse/IDEA can connect).

Note that this change will not affect production, as the JVM debugger is only
enabled via the local working copy of docker-compose.yml and the ./run script,
which are strictly intended for developer use only (Ross writes his own
docker-compose.yml files for managing microservices in production, which leave
JVM debugging disabled by default).