Configure the application to use JSON formatted logs.
This relies on a dgit library for formatting.  The `log4j2.properties` file defines a configuration that logs JSON to standard out.  It's based on the one used in Telflow karaf applications.

This logger and configuration is available for both gradle- and maven-built applications.

The template does its best to force all logging through slf4j and thence to log4j2.

Also, clean up some gradle deprecation warnings.